+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
+
+## [0.8.3] — 2024-09-02
 * Add streams/with-out-stream
 * Add base64/encode-str
 * Bump version of clojure/core.async
@@ -59,4 +61,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [0.8.0]: https://github.com/tanelso2/clj-toolbox/compare/0.7.3...0.8.0
 [0.8.1]: https://github.com/tanelso2/clj-toolbox/compare/0.8.0...0.8.1
 [0.8.2]: https://github.com/tanelso2/clj-toolbox/compare/0.8.1...0.8.2
-[Unreleased]: https://github.com/tanelso2/clj-toolbox/compare/0.8.2...HEAD
+[0.8.3]: https://github.com/tanelso2/clj-toolbox/compare/0.8.2...0.8.3
+[Unreleased]: https://github.com/tanelso2/clj-toolbox/compare/0.8.3...HEAD